requirements.txt
Django>=5.0.6,<5.1
djangorestframework>=3.15.2,<3.16
django-filter==24.2
dj-rest-auth[with_social]>=6.0.0,<6.1
djangorestframework-simplejwt>=5.3.1,<5.4
drf-spectacular>=0.26.5,<0.27
psycopg2>=2.9.9,<3
yookassa>=3.3.0,<3.4
pillow>=10.4.0,<10.5
# TODO: Install pillow!